# backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/hr/attendance/check-in": {
      "post": {
        "operationId": "Attendance_checkIn",
        "summary": "Check in for today",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/hr/attendance/check-out": {
      "post": {
        "operationId": "Attendance_checkOut",
        "summary": "Check out for today",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/hr/attendance": {
      "get": {
        "operationId": "Attendance_findAll",
        "summary": "Get all attendance records (HR/Manager/Admin)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AttendanceResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/hr/attendance/me": {
      "get": {
        "operationId": "Attendance_getMyAttendance",
        "summary": "Get current user attendance records",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttendanceResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/api/v1/hr/salary-slips": {
      "post": {
        "operationId": "Salary_create",
        "summary": "Create salary slip",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSalarySlipDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalarySlipResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "Salary_findAll",
        "summary": "Get all salary slips",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isPaid",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SalarySlipResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/hr/salary-slips/me": {
      "get": {
        "operationId": "Salary_getMySlips",
        "summary": "Get my salary slips",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalarySlipResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/hr/salary-slips/{id}/pay": {
      "patch": {
        "operationId": "Salary_markAsPaid",
        "summary": "Mark salary slip as paid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalarySlipResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TeaP ERP/POS API",
    "description": "API documentation for TeaP — Bubble Tea Chain ERP/POS Management System. Supports multi-branch operations, POS, inventory management, HR, and finance.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT access token",
        "in": "header"
      }
    },
    "schemas": {
...
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {}
      },
...
      "CheckInDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "description": "Note for check-in or check-out"
          }
        }
      },
      "HrUserReferenceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "branchId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "fullName"
        ]
      },
      "AttendanceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "checkIn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "checkOut": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "hoursWorked": {
            "type": "string",
            "nullable": true,
            "example": "7.5",
            "pattern": "^-?\\d+(?:\\.\\d+)?$",
            "description": "Exact decimal hours"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/HrUserReferenceDto"
          }
        },
        "required": [
          "id",
          "userId",
          "date",
          "createdAt"
        ]
      },
...
      "CreateSalarySlipDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "month": {
            "type": "number"
          },
          "year": {
            "type": "number"
          },
          "baseSalary": {
            "type": "number"
          },
          "bonus": {
            "type": "number"
          },
          "deduction": {
            "type": "number"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "month",
          "year",
          "baseSalary"
        ]
      },
      "SalarySlipResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "month": {
            "type": "number",
            "minimum": 1,
            "maximum": 12
          },
          "year": {
            "type": "number"
          },
          "baseSalary": {
            "type": "integer",
            "description": "Whole VND"
          },
          "bonus": {
            "type": "integer",
            "description": "Whole VND"
          },
          "deduction": {
            "type": "integer",
            "description": "Whole VND"
          },
          "netSalary": {
            "type": "integer",
            "description": "Whole VND"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "isPaid": {
            "type": "boolean"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/HrUserReferenceDto"
          }
        },
        "required": [
          "id",
          "userId",
          "month",
          "year",
          "baseSalary",
          "bonus",
          "deduction",
          "netSalary",
          "isPaid",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}